requirements-prd.txt
dispatch==0.3.12
git+git://github.com/ubyssey/django-google-storage.git
django-gae-backends
django-canonical-domain
mysql-python
django-phonenumber-field
beautifulsoup4
